Fix cmp peek and poke csp node length

+++ include/csp/csp_cmp.h
@@ -194,7 +194,7 @@ CMP_MESSAGE(CSP_CMP_CLOCK, clock)
    @param[out] msg peeked/read memory.
    @return #CSP_ERR_NONE on success, otherwise an error code.
 */
-static inline int csp_cmp_peek(uint8_t node, uint32_t timeout, struct csp_cmp_message *msg) {
+static inline int csp_cmp_peek(uint16_t node, uint32_t timeout, struct csp_cmp_message *msg) {
     return csp_cmp(node, timeout, CSP_CMP_PEEK, CMP_SIZE(peek) - sizeof(msg->peek.data) + msg->peek.len, msg);
 }
 
@@ -205,7 +205,7 @@ static inline int csp_cmp_peek(uint8_t node, uint32_t timeout, struct csp_cmp_me
    @param[in] msg memory address, number of bytes and the actual bytes to poke/write.
    @return #CSP_ERR_NONE on success, otherwise an error code.
 */
-static inline int csp_cmp_poke(uint8_t node, uint32_t timeout, struct csp_cmp_message *msg) {
+static inline int csp_cmp_poke(uint16_t node, uint32_t timeout, struct csp_cmp_message *msg) {
     return csp_cmp(node, timeout, CSP_CMP_POKE, CMP_SIZE(poke) - sizeof(msg->poke.data) + msg->poke.len, msg);
 }
